Us 24

+++ src/main/java/ar/edu/uner/prestabook/model/CodigoIdentificatorio.java
@@ -17,13 +17,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CodigoIdentificatorio {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(unique = true, nullable = false)
-    private Long id;
-    private Integer pasillo;
-    private Integer estanteria;
-    private Integer estante;
+	 	@Id
+	    @GeneratedValue(strategy = GenerationType.AUTO)
+	    @Column(unique = true, nullable = false)
+	    private Integer id;
+	    private Integer pasillo;
+	    private Integer estanteria;
+	    private Integer estante;
 
 }
